requirements.txt
tensorflow-gpu==2.0.0a0
tensorboard==2.1.0
numpy==1.18.0
librosa==0.7.2
argparse
scipy==1.4.1
tqdm
glob2
matplotlib
sounddevice
